requirements.txt
pymavlink
opencv-python
pyserial
rplidar